refactor: move handler to handler.go
