+++ aftermath/internal/lsp/handler.go
@@ -1,0 +1,52 @@
+package lsp
+
+import (
+	"log"
+
+	"github.com/tliron/glsp"
+	protocol "github.com/tliron/glsp/protocol_3_16"
+)
+
+func (ls *LanguageServer) initialize(
+	context *glsp.Context,
+	params *protocol.InitializeParams,
+) (any, error) {
+	capabilities := ls.handler.CreateServerCapabilities()
+
+	return protocol.InitializeResult{
+		Capabilities: capabilities,
+		ServerInfo: &protocol.InitializeResultServerInfo{
+			Name:    lsName,
+			Version: &version,
+		},
+	}, nil
+}
+
+func (ls *LanguageServer) initialized(
+	context *glsp.Context,
+	params *protocol.InitializedParams,
+) error {
+	log.Println("Server initialized")
+	return nil
+}
+
+func (ls *LanguageServer) shutdown(context *glsp.Context) error {
+	log.Println("Server shutting down")
+	protocol.SetTraceValue(protocol.TraceValueOff)
+	return nil
+}
+
+func (ls *LanguageServer) setTrace(context *glsp.Context, params *protocol.SetTraceParams) error {
+	protocol.SetTraceValue(params.Value)
+	log.Printf("Trace set to: %s", params.Value)
+	return nil
+}
+
+func (ls *LanguageServer) executeCommand(
+	context *glsp.Context,
+	params *protocol.ExecuteCommandParams,
+) (interface{}, error) {
+	// log.Printf("State is %s", ls.state)
+	log.Println("Hello World")
+	return nil, nil
+}
